# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ApiKeyDto": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "ApiKeyDto",
        "type": "object"
      },
      "GetTimesResDto": {
        "properties": {
          "times": {
            "items": {
              "$ref": "#/components/schemas/GetTimesResDto.TimeDto"
            },
            "title": "Times",
            "type": "array"
          }
        },
        "required": [
          "times"
        ],
        "title": "GetTimesResDto",
        "type": "object"
      },
      "GetTimesResDto.TimeDto": {
        "properties": {
          "cache_file": {
            "title": "Cache File",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "millis": {
            "title": "Millis",
            "type": "integer"
          },
          "output_json": {
            "title": "Output Json",
            "type": "string"
          },
          "poster": {
            "$ref": "#/components/schemas/GetTimesResDto.UserDto"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "trace_file": {
            "title": "Trace File",
            "type": "string"
          },
          "validated": {
            "title": "Validated",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "millis",
          "cache_file",
          "trace_file",
          "validated",
          "output_json",
          "poster",
          "language",
          "timestamp"
        ],
        "title": "TimeDto",
        "type": "object"
      },
      "GetTimesResDto.UserDto": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_verified"
        ],
        "title": "UserDto",
        "type": "object"
      },
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LoginDto": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginDto",
        "type": "object"
      },
      "RecordTimeReqDto": {
        "properties": {
          "cache_file": {
            "title": "Cache File",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "millis": {
            "title": "Millis",
            "type": "integer"
          },
          "output_json": {
            "title": "Output Json",
            "type": "string"
          },
          "trace_file": {
            "title": "Trace File",
            "type": "string"
          }
        },
        "required": [
          "millis",
          "cache_file",
          "trace_file",
          "output_json",
          "language"
        ],
        "title": "RecordTimeReqDto",
        "type": "object"
      },
      "SignupDto": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "minLength": 6,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "SignupDto",
        "type": "object"
      },
      "TimeDto": {
        "properties": {
          "cache_file": {
            "title": "Cache File",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "millis": {
            "title": "Millis",
            "type": "integer"
          },
          "output_json": {
            "title": "Output Json",
            "type": "string"
          },
          "poster": {
            "$ref": "#/components/schemas/TimeDto.UserDto"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "trace_file": {
            "title": "Trace File",
            "type": "string"
          },
          "validated": {
            "title": "Validated",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "millis",
          "cache_file",
          "trace_file",
          "validated",
          "output_json",
          "poster",
          "language",
          "timestamp"
        ],
        "title": "TimeDto",
        "type": "object"
      },
      "TimeDto.UserDto": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_verified"
        ],
        "title": "UserDto",
        "type": "object"
      },
      "UserDto": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_verified"
        ],
        "title": "UserDto",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "detail": {
            "properties": {
              "<location>": {
                "properties": {
                  "<field_name>": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "VerifyDto": {
        "properties": {
          "otp": {
            "title": "Otp",
            "type": "string"
          }
        },
        "required": [
          "otp"
        ],
        "title": "VerifyDto",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "CookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "arch-leader",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/auth/api-key": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyDto"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "CookieAuth": []
          }
        ],
        "summary": "Api Key",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "CookieAuth": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/resend-otp": {
      "patch": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "CookieAuth": []
          }
        ],
        "summary": "Resend Otp",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Signup",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify": {
      "patch": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "CookieAuth": []
          }
        ],
        "summary": "Verify",
        "tags": [
          "Auth"
        ]
      }
    },
    "/heartbeat/": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "Heartbeat"
        ]
      }
    },
    "/time/": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTimesResDto"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get All",
        "tags": [
          "Time"
        ]
      }
    },
    "/time/record": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordTimeReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeDto"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "CookieAuth": []
          }
        ],
        "summary": "Record",
        "tags": [
          "Time"
        ]
      }
    },
    "/user/": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "CookieAuth": []
          }
        ],
        "summary": "Get Current User",
        "tags": [
          "User"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "tags": [
    {
      "name": "User"
    },
    {
      "name": "Auth"
    },
    {
      "name": "Time"
    },
    {
      "name": "Heartbeat"
    }
  ]
}